Refactor search agent after SQL retriever removal
